We've really got an actual game now
Build out an interesting first level, mess around with prefabs a bit,
import a fuckton of meshes, set push to left stick, add a gravity
button, lots of other little tweaks.

The Gravity button needs to be stronger, it's fairly subtle now. We also
have the machinery in place to change the direction of gravity, if we
want to.